requirements.txt
# ============================================
# RoboMarty Requirements
# Install with: pip install -r requirements.txt
# ============================================

# Core Robot Control
martypy==2.1.0

# Computer Vision
opencv-python==4.8.1.78
numpy==1.24.3

# Deep Learning Frameworks
tensorflow==2.15.0
tf-keras==2.15.0

# Facial Emotion Detection
deepface==0.0.79
mtcnn==0.1.1
retina-face==0.0.13

# Object Detection (YOLO)
ultralytics==8.0.196
torch==2.1.0
torchvision==0.16.0

# Image Processing
Pillow==10.1.0

# Speech Recognition
SpeechRecognition==3.10.0
PyAudio==0.2.14

# OpenAI API (Whisper & GPT)
openai==1.3.5

# Text-to-Speech
elevenlabs==0.2.27

# Audio Processing & Playback
sounddevice==0.4.6
scipy==1.11.3

# Environment Variables
python-dotenv==1.0.0

# Mathematics
sympy==1.12

# Additional Dependencies
pandas==2.1.3
requests==2.31.0
urllib3==2.1.0
protobuf==4.25.1
charset-normalizer==3.3.2

# Data Science & ML Support
scikit-learn==1.3.2
matplotlib==3.8.1
seaborn==0.13.0

# Async & Threading (built-in but dependencies)
typing-extensions==4.8.0

# File Operations
gdown==4.7.1

# Additional ML Backend Support
onnxruntime==1.16.3

# ============================================
# INSTALLATION INSTRUCTIONS
# ============================================
#
# 1. Create virtual environment:
#    python -m venv venv
#
# 2. Activate virtual environment:
#    Windows:   venv\Scripts\activate
#    Mac/Linux: source venv/bin/activate
#
# 3. Install all requirements:
#    pip install -r requirements.txt
#
# ============================================
# PLATFORM-SPECIFIC NOTES
# ============================================
#
# WINDOWS:
# - PyAudio should install directly via pip
# - If issues, download wheel from:
#   https://www.lfd.uci.edu/~gohlke/pythonlibs/#pyaudio
#   Then: pip install PyAudio‑0.2.11‑cp3X‑cp3Xm‑win_amd64.whl
#
# MAC:
# - Install portaudio first: brew install portaudio
# - Then: pip install pyaudio
#
# LINUX (Ubuntu/Debian):
# - sudo apt-get update
# - sudo apt-get install python3-pyaudio portaudio19-dev
# - sudo apt-get install python3-dev
# - pip install pyaudio
#
# ============================================
# GPU ACCELERATION (OPTIONAL)
# ============================================
#
# For NVIDIA GPU support (faster processing):
#
# Instead of torch==2.1.0, install CUDA version:
# pip install torch==2.1.0+cu118 torchvision==0.16.0+cu118 --index-url https://download.pytorch.org/whl/cu118
#
# For TensorFlow GPU:
# pip install tensorflow[and-cuda]==2.15.0
#
# ============================================
# TROUBLESHOOTING
# ============================================
#
# If you get version conflicts:
# pip install --upgrade pip
# pip install -r requirements.txt --upgrade
#
# If specific package fails:
# pip install package-name --no-deps
# Then reinstall: pip install -r requirements.txt
#
# For clean reinstall:
# pip uninstall -r requirements.txt -y
# pip install -r requirements.txt
#
# ============================================